# src/l1/piccolod/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Piccolo OS Admin API (Draft)
  version: 0.1.0
security:
  - cookieAuth: []
servers:
  - url: /api/v1
paths:
...
  /crypto/setup:
    post:
      summary: Initialize crypto using admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /crypto/unlock:
    post:
      summary: Unlock crypto with admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
                recovery_key: { type: string, description: Space-separated 24-word mnemonic }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        details: { type: string, nullable: true }
        next_step: { type: string, nullable: true }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: piccolo_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token